feat: 모달 컴포넌트 개발 (#20)
* feat: 상품 클릭시 모달 열기, 담기 버튼 클릭 시 모달 닫기

* design: Modal 개선

* design: Modal 화면에 꽉 차게 개선

* feat: 모달 Context 생성 및 적용

* feat: 결제방식선택화면 - 모달 연결

* feat: 로딩화면 - 모달 연결

* feat: 장바구니에서 결제방식 선택 모달 열기 기능

* feat: 현금결제화면에서 영수증모달 열기 기능

* feat: 영수증 화면 닫기 기능

* feat: 메뉴 클릭 시 메뉴 담기 모달 열기 기능

* style: 옵션 선택 화면의 선택된 메뉴 업데이트 함수명 변경

* refactor: 불필요한 함수 제거

+++ packages/client/src/components/main/Receipt.tsx
@@ -1,4 +1,5 @@
-import { useEffect, useState } from 'react'
+import { ModalContext } from 'App'
+import { useContext, useEffect, useState } from 'react'
 import { OrderType, SelectedMenuType } from 'type'
 
 interface ReceiptProps {
@@ -45,13 +46,15 @@ function OrderedMenuList({ selectedMenuList }: OrderedMenuListProps) {
 }
 
 function Timer({ deleteAllCartMenu }: TimerProps) {
+  const [isModalOpen, setIsModalOpen] = useContext(ModalContext)
+
   const INITIAL_LEFT_TIME = 10
   const [leftTime, setLeftTime] = useState(INITIAL_LEFT_TIME)
 
   useEffect(() => {
     const interval = setInterval(() => {
       setLeftTime((leftTIme) => leftTIme - 1)
-    }, 1000)
+    }, 300)
 
     return () => {
       clearInterval(interval)
@@ -60,8 +63,9 @@ function Timer({ deleteAllCartMenu }: TimerProps) {
 
   useEffect(() => {
     if (leftTime === 0) {
+      if (!isModalOpen || !setIsModalOpen) return
+      setIsModalOpen({ ...isModalOpen, receipt: false })
       deleteAllCartMenu()
-      console.log('close receipt')
     }
   }, [leftTime])
 
@@ -81,7 +85,6 @@ export default function Receipt({
 
   return (
     <div>
-      <button onClick={deleteAllCartMenu}>영수증 닫기</button>
       <div>주문 번호 {orderNumber}</div>
       <OrderedMenuList selectedMenuList={selectedMenuList} />
       <div>
